refactor: changed names and remove desnecessary add method
